FEATURE : 오늘의 옷 추천 기능 구현과 게시글, 옷이미지 삭제시 확인 알람 추가

+++ client/src/components/community/MarketViewForm.js
@@ -147,11 +147,19 @@ const MarketViewForm = ({ postType }) => {
   };
 
   const handleRemoveButton = async () => {
+    if (!window.confirm("게시글을 삭제하시겠습니까?")) {
+      alert("취소 했습니다..");
+      return;
+    }
+
     try {
       await axios.delete(port.url + `/api/market/list/${paramsId}/delete`, {
         headers: { accessToken: cookies.userData.accessToken },
       });
     } catch {}
+
+    alert("게시글을 삭제했습니다.");
+
     navigate("/market");
     window.location.reload();
   };
@@ -236,7 +244,9 @@ const MarketViewForm = ({ postType }) => {
                     right: 0,
                   }}
                 >
-                  조회수 : {curPost && curPost.views}
+                  작성일 : {curPost && curPost.createdAt.split("T")[0]}
+                  {" / "}
+                  {"      "} 조회수 : {curPost && curPost.views}
                 </div>
               </ContentRowDivide>
             </ContentRow>
